refactor(backend): improve code quality

+++ backend/src/course/types.ts
@@ -4,7 +4,7 @@ import { z } from "zod";
 export const courseSchema = z.object({
   id: z.number().int().positive(),
   name: z.string().nonempty(),
-  maxScore: z.number().positive(),
+  maxScore: z.number().int().positive(),
 });
 
 export type Course = Infer<typeof courseSchema>;
